requirements.txt
streamlit
requests
beautifulsoup4
faiss-cpu
numpy
fastembed
transformers
torch
huggingface-hub
python-dotenv